Add logging module
Logging is an important feature for most applications, especially for
retroactively diagnosing problems. The Argent Middleware Library
provides support for logging through the following interface of the
logging module:
  - `agm_log_open()` function
  - `agm_log_close()` function
  - `agm_log_trace()` macro
  - `agm_log_warning()` macro
  - `agm_log_error()` macro
  - `agm_log_erno()` macro

The namespace has also been changed to `agm_` from `arm_`, reflecting
the change in the namespace of the Argent Core Library from `arc_` to
`agc_`.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>